backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
cryptography==41.0.7
python-multipart==0.0.6
